Make preference of user vs. alias in address-resolution configurable
Before, it was hardwired to always prefer aliases over user-accounts on a
domain. This may lead to problematic situation, e.g. when you want a
full-catchall on a domain, but still be able to direct single, specific
addresses elsewhere.

Here, we make the order of resolution configurable, keeping the default on the
current behavior.

closes #815